[ch #159961943] Move app component into components
